+++ apps/exam-app/src/components/exam/styles.ts
@@ -1,0 +1,331 @@
+import { Platform } from "react-native";
+import { StyleSheet } from "react-native-unistyles";
+
+export const styles = StyleSheet.create((theme) => ({
+  container: {
+    flex: 1,
+    backgroundColor: theme.colors.background,
+  },
+  center: {
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  header: {
+    paddingTop: Platform.OS === "android" ? 35 : 15,
+    paddingBottom: 12,
+    paddingHorizontal: 16,
+    backgroundColor: theme.colors.surface,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    borderBottomWidth: 1,
+    borderBottomColor: theme.colors.border,
+    elevation: 2,
+    zIndex: 10,
+  },
+  headerTitleContainer: {
+    flex: 1,
+    marginRight: 10,
+  },
+  examTitle: {
+    fontSize: 16,
+    fontWeight: "bold",
+    color: theme.colors.typography,
+  },
+  headerActions: {
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 8,
+  },
+  iconButton: {
+    padding: 8,
+    borderRadius: 20,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  timerContainer: {
+    paddingHorizontal: 10,
+    paddingVertical: 6,
+    backgroundColor: theme.colors.inputBg,
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+  },
+  timerText: {
+    fontSize: 14,
+    fontWeight: "bold",
+    fontVariant: ["tabular-nums"],
+    color: theme.colors.typography,
+  },
+  warningBadge: {
+    flexDirection: "row",
+    alignItems: "center",
+    paddingHorizontal: 8,
+    paddingVertical: 6,
+    borderRadius: 8,
+    gap: 4,
+  },
+  warningText: {
+    color: "#fff",
+    fontSize: 12,
+    fontWeight: "bold",
+  },
+  contentContainer: {
+    padding: 16,
+    paddingBottom: 100,
+  },
+  questionCard: {
+    backgroundColor: theme.colors.surface,
+    borderRadius: 12,
+    padding: 20,
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.05,
+    shadowRadius: 4,
+    elevation: 2,
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+    minHeight: 400,
+  },
+  questionHeader: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+    marginBottom: 20,
+  },
+  questionNumber: {
+    fontSize: 18,
+    fontWeight: "bold",
+    color: theme.colors.typography,
+  },
+  typeBadge: {
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 6,
+  },
+  htmlContainer: {
+    marginBottom: 24,
+  },
+  divider: {
+    height: 1,
+    backgroundColor: theme.colors.border,
+    marginBottom: 24,
+  },
+  dividerSuccess: {
+    height: 1,
+    backgroundColor: theme.colors.border,
+  },
+  answerArea: {
+    gap: 16,
+  },
+  radioOption: {
+    flexDirection: "row",
+    alignItems: "flex-start",
+    padding: 16,
+    borderRadius: 8,
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+    backgroundColor: theme.colors.background,
+  },
+  radioOptionSelected: {
+    backgroundColor: theme.colors.inputBg,
+    borderColor: theme.colors.primary,
+    borderWidth: 1.5,
+  },
+  radioCircle: {
+    width: 22,
+    height: 22,
+    borderRadius: 11,
+    borderWidth: 2,
+    borderColor: theme.colors.muted,
+    marginRight: 16,
+    marginTop: 2,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  radioCircleSelected: {
+    borderColor: theme.colors.primary,
+  },
+  radioDot: {
+    width: 12,
+    height: 12,
+    borderRadius: 6,
+    backgroundColor: theme.colors.primary,
+  },
+  essayInput: {
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+    borderRadius: 8,
+    padding: 16,
+    minHeight: 200,
+    textAlignVertical: "top",
+    fontSize: 16,
+    backgroundColor: theme.colors.background,
+    color: theme.colors.typography,
+  },
+  footer: {
+    position: "absolute",
+    bottom: 0,
+    left: 0,
+    right: 0,
+    height: 80,
+    backgroundColor: theme.colors.surface,
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between",
+    paddingHorizontal: 24,
+    borderTopWidth: 1,
+    borderTopColor: theme.colors.border,
+    elevation: 10,
+  },
+  navButton: {
+    width: 48,
+    height: 48,
+    borderRadius: 24,
+    backgroundColor: theme.colors.inputBg,
+    justifyContent: "center",
+    alignItems: "center",
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+  },
+  navButtonDisabled: {
+    opacity: 0.3,
+  },
+  sheetButton: {
+    flexDirection: "row",
+    alignItems: "center",
+    backgroundColor: theme.colors.inputBg,
+    paddingVertical: 10,
+    paddingHorizontal: 16,
+    borderRadius: 20,
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+    gap: 8,
+  },
+  sheetButtonText: {
+    fontSize: 14,
+    fontWeight: "600",
+    color: theme.colors.typography,
+  },
+  gridIcon: {
+    flexDirection: "row",
+    flexWrap: "wrap",
+    width: 14,
+    height: 14,
+    gap: 2,
+    justifyContent: "center",
+  },
+  gridDot: {
+    width: 5,
+    height: 5,
+    backgroundColor: theme.colors.typography,
+    borderRadius: 1,
+  },
+  finishButton:
+    Platform.OS === "web"
+      ? {
+          backgroundColor: theme.colors.primary,
+          height: 48,
+          paddingHorizontal: 24,
+          borderRadius: 24,
+          justifyContent: "center",
+          alignItems: "center",
+          shadowColor: theme.colors.primary,
+          shadowOffset: { width: 0, height: 4 },
+          shadowOpacity: 0.3,
+          shadowRadius: 4,
+          elevation: 4,
+        }
+      : {
+          backgroundColor: theme.colors.primary,
+          height: 48,
+          paddingHorizontal: 24,
+          borderRadius: 24,
+          justifyContent: "center",
+          alignItems: "center",
+        },
+  finishButtonText: {
+    color: "#fff",
+    fontWeight: "bold",
+    fontSize: 16,
+  },
+  gridContainer: {
+    flexDirection: "row",
+    flexWrap: "wrap",
+    gap: 12,
+    justifyContent: "center",
+    paddingVertical: 16,
+  },
+  gridItem: {
+    width: 50,
+    height: 50,
+    justifyContent: "center",
+    alignItems: "center",
+    borderRadius: 8,
+  },
+  modalCloseButton: {
+    backgroundColor: theme.colors.surface,
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    borderRadius: 8,
+  },
+  modalCloseText: {
+    color: theme.colors.typography,
+    fontWeight: "bold",
+  },
+  buttonPrimary: {
+    backgroundColor: theme.colors.primary,
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    borderRadius: 8,
+  },
+  buttonText: { color: "#fff", fontWeight: "bold" },
+  modalFooter: {
+    flexDirection: "row",
+    justifyContent: "flex-end",
+    gap: 12,
+    width: "100%",
+  },
+  modalButton: {
+    paddingVertical: 10,
+    paddingHorizontal: 16,
+    borderRadius: 8,
+  },
+  modalButtonDefault: {
+    backgroundColor: theme.colors.primary,
+  },
+  modalButtonCancel: {
+    backgroundColor: "transparent",
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+  },
+  modalButtonText: {
+    fontWeight: "bold",
+    color: "#fff",
+  },
+  // Info Card untuk Sukses Page
+  infoCard: {
+    width: "100%",
+    backgroundColor: theme.colors.surface,
+    borderRadius: 12,
+    padding: 16,
+    borderWidth: 1,
+    borderColor: theme.colors.border,
+    gap: 12,
+  },
+  infoRow: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  screenshotAlert: {
+    marginTop: 24,
+    padding: 12,
+    backgroundColor: theme.colors.inputBg,
+    borderRadius: 8,
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 10,
+  },
+}));
